Added Launch Party Icon (#98)
* Launch Party Prefab Setup

This commit includes all the necessary things to get started with Launch Party (base color sprites, prefab, background, and a loader class script). Anyone is welcome to finish it

* Launch Party Icon Added

Due to me being stupid, I forgot to put the icon in the existing pull request. This one should fix that