HD Fork Lifter (#292)
* starting out with double date stuff :D

not even the background is finished
i just wanna get this on my fork so that it's safe

* double date getting more initialized

no animations, one block, nothing actually functions. but the boy is put in place, and the girl is almost put in place! just wanted to merge this with the main branch to play catchy tune

* initialization done!!!!!

gonna fix up the code, see what i can take out, see what i can standardize, see what i need to add. loving this so far, even with all of its annoyances

* ughhhh animation stuff.

this is gonna take me a day or two to even comprehend

* starting on animation

committing cause im traveling, i might work on this if i have time

* replace sounds, extend some stuff

replaced the rhre3 sounds and added a few things to make camera movements not look so out-of-bounds janky. still working on it, but it's coming along good so far

* starting on dog ninja

waiting for bread2unity to get merged, and i'm still probably gonna keep working on dog ninja until it's done even if it does.

* initialization and sprite trimming done

reference is from angrytapper2's video

Throw Object,
Cut Everything!, and
Here We Go!

were the cues added, in that order.

* temp icon w/ mask + bezier curves!!

the icon is temporary, but i did add a mask. i think it looks pretty good :) especially for  what it is. also bezier curves. i just figured those out, now i just need to figure out how to implement them into the object.

* getting there

i'm so close to getting the "Cut Everything!" block to work. i'm trying to figure out how to not play it as the default when the game is rendered

* added sfx
* added bird animations
* looked at double date for like 2 minutes
* have been editing blue bear's "treat" block to fit dog ninja for a while now

(im gonna try doing these commit notes from now on)

* keeping trying

* cues have sfx now
 - the "throw object" cues are not accurate yet
* animations adjusted to not keep bird on screen

* not much progress rn, getting into the harder stuff

* added more sfx
 - it's from rhre3, if somebody can find a clean rip that would be great, cause i can't)
* made a bop animation
 - still trying to figure out how to get it to play every beat :/
* replaced throw sfx, was using whiff sfx before
* extended background using a slice of the sky gradient and color picked blocks
* some code added to dogninja.cs and throwobject.cs is... bad. it's just torn from other games.

* i need to work out conflicts

really this is just pushing all my stuff so i don't lose it, tambourine is giving my fork conflicts for some reason :skull:

* blehhh

figuring out git some more

* fork lifter hd sprites

courtesy of unbrokenmage

* had to remove more files

* had to add files!

* oops! more changes!

i gotta add the little hand now... people are asking me to add it...

* i cannot do the little hand :(

* put the two background elements into one group
* tweaked the position of the hand visual (it wasn't pointing right where the peas were coming from)

---------

Co-authored-by: minenice55 <star.elementa@gmail.com>
Co-authored-by: patataofcourse <patataofcourse@gmail.com>