Mr. Upbeat Rework (#418)
* everything but the text

* new spritesheet (thank u sean)
* better/more accurate background color
* bg color changing with fade (code courtesy of ras)
* mr. upbeat blip color changing
* full blip implementation, with size changing and text display

* text is still weird (gray box around all the letters) but i hope that's an easy fix.

* almost done

just need to fix the text and get rid of the reliance on the ding! block

* finishing touches + a buncha testing

everything works!!! and i don't think i left anything out

* forgot an n lol

* oh my god i hate mr. downbeat

this "force" option is not very intuitive to use. nor is it coded amazingly. but i do not care. if you really wanna use it then mess around and see what works